+++ Sources/VCore/Models/Edge Insets/EdgeInsets_LeadingTrailingTopBottom.swift
@@ -171,10 +171,25 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
     }
 
 #endif
-    
-    // MARK: Insetting
-    /// Insets `EdgeInsets` by a given value.
-    public func insetBy(inset: CGFloat) -> EdgeInsets_LeadingTrailingTopBottom {
+
+    // MARK: Map
+    /// Returns `EdgeInsets_LeadingTrailingTopBottom`  containing the results of mapping the given closure over the values.
+    public func map(
+        _ transform: (CGFloat) throws -> CGFloat
+    ) rethrows -> Self {
+        .init(
+            leading: try transform(leading),
+            trailing: try transform(trailing),
+            top: try transform(top),
+            bottom: try transform(bottom)
+        )
+    }
+
+    // MARK: Inset
+    /// Insets `EdgeInsets_LeadingTrailingTopBottom` by a given value.
+    public func insetBy(
+        inset: CGFloat
+    ) -> EdgeInsets_LeadingTrailingTopBottom {
         .init(
             leading: leading + inset,
             trailing: trailing + inset,
@@ -183,8 +198,11 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
         )
     }
     
-    /// Insets `EdgeInsets` by a given horizontal and vertical values.
-    public func insetBy(horizontal horizontalInset: CGFloat, vertical verticalInset: CGFloat) -> EdgeInsets_LeadingTrailingTopBottom {
+    /// Insets `EdgeInsets_LeadingTrailingTopBottom` by a given horizontal and vertical values.
+    public func insetBy(
+        horizontal horizontalInset: CGFloat,
+        vertical verticalInset: CGFloat
+    ) -> EdgeInsets_LeadingTrailingTopBottom {
         .init(
             leading: leading + horizontalInset,
             trailing: trailing + horizontalInset,
@@ -193,8 +211,10 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
         )
     }
     
-    /// Insets `EdgeInsets` by a given leading value.
-    public func insetBy(leading leadingInset: CGFloat) -> EdgeInsets_LeadingTrailingTopBottom {
+    /// Insets `EdgeInsets_LeadingTrailingTopBottom` by a given leading value.
+    public func insetBy(
+        leading leadingInset: CGFloat
+    ) -> EdgeInsets_LeadingTrailingTopBottom {
         .init(
             leading: leading + leadingInset,
             trailing: trailing,
@@ -203,8 +223,10 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
         )
     }
     
-    /// Insets `EdgeInsets` by a given trailing value.
-    public func insetBy(trailing trailingInset: CGFloat) -> EdgeInsets_LeadingTrailingTopBottom {
+    /// Insets `EdgeInsets_LeadingTrailingTopBottom` by a given trailing value.
+    public func insetBy(
+        trailing trailingInset: CGFloat
+    ) -> EdgeInsets_LeadingTrailingTopBottom {
         .init(
             leading: leading,
             trailing: trailing + trailingInset,
@@ -213,8 +235,10 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
         )
     }
     
-    /// Insets `EdgeInsets` by a given top value.
-    public func insetBy(top topInset: CGFloat) -> EdgeInsets_LeadingTrailingTopBottom {
+    /// Insets `EdgeInsets_LeadingTrailingTopBottom` by a given top value.
+    public func insetBy(
+        top topInset: CGFloat
+    ) -> EdgeInsets_LeadingTrailingTopBottom {
         .init(
             leading: leading,
             trailing: trailing,
@@ -223,8 +247,10 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
         )
     }
     
-    /// Insets `EdgeInsets` by a given bottom value.
-    public func insetBy(bottom bottomInset: CGFloat) -> EdgeInsets_LeadingTrailingTopBottom {
+    /// Insets `EdgeInsets_LeadingTrailingTopBottom` by a given bottom value.
+    public func insetBy(
+        bottom bottomInset: CGFloat
+    ) -> EdgeInsets_LeadingTrailingTopBottom {
         .init(
             leading: leading,
             trailing: trailing,
@@ -234,7 +260,7 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
     }
     
     // MARK: Operators
-    /// Adds two `EdgeInsets` by adding up individual edge insets.
+    /// Adds two `EdgeInsets_LeadingTrailingTopBottom` by adding up individual edge insets.
     public static func + (lhs: Self, rhs: Self) -> Self {
         .init(
             leading: lhs.leading + rhs.leading,
@@ -244,7 +270,7 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
         )
     }
     
-    /// Adds right `EdgeInsets` to the left one by adding individual edge insets.
+    /// Adds right `EdgeInsets_LeadingTrailingTopBottom` to the left one by adding individual edge insets.
     public static func += (lhs: inout Self, rhs: Self) {
         lhs.leading += rhs.leading
         lhs.trailing += rhs.trailing
@@ -252,7 +278,7 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
         lhs.bottom += rhs.bottom
     }
     
-    /// Subtracts two `EdgeInsets` by subtracting up individual edge insets.
+    /// Subtracts two `EdgeInsets_LeadingTrailingTopBottom` by subtracting up individual edge insets.
     public static func - (lhs: Self, rhs: Self) -> Self {
         .init(
             leading: lhs.leading - rhs.leading,
@@ -262,7 +288,7 @@ public struct EdgeInsets_LeadingTrailingTopBottom: Equatable, Hashable {
         )
     }
     
-    /// Subtracts right `EdgeInsets` to the left one by subtracting individual edge insets.
+    /// Subtracts right `EdgeInsets_LeadingTrailingTopBottom` to the left one by subtracting individual edge insets.
     public static func -= (lhs: inout Self, rhs: Self) {
         lhs.leading -= rhs.leading
         lhs.trailing -= rhs.trailing