+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Views
 
 Models
 
+- `map(_:)` method is added to `EdgeInsets`, that returns itself containing the results of mapping the given closure over the values.
 - `map(_:)` method is added to `GenericStateModels`, that returns itself containing the results of mapping the given closure over the values.
 
 Extensions